add SupportedChainIds to types
