+++ packages/swapper/src/swappers/zrx/utils/helpers/helpers.ts
@@ -1,12 +1,12 @@
 import { ChainAdapterManager } from '@shapeshiftoss/chain-adapters'
 import { HDWallet } from '@shapeshiftoss/hdwallet-core'
-import { Asset, QuoteResponse } from '@shapeshiftoss/types'
+import { Asset, SupportedChainIds, QuoteResponse } from '@shapeshiftoss/types'
 import { AxiosResponse } from 'axios'
 import BigNumber from 'bignumber.js'
 import Web3 from 'web3'
 import { AbiItem, numberToHex } from 'web3-utils'
 
-import { SupportedChainIds, SwapError, TradeQuote } from '../../../../api'
+import { SwapError, TradeQuote } from '../../../../api'
 import { ZrxError } from '../../ZrxSwapper'
 import { bnOrZero } from '../bignumber'
 import { zrxService } from '../zrxService'
