+++ packages/swapper/src/swappers/zrx/zrxBuildTrade/zrxBuildTrade.ts
@@ -1,7 +1,8 @@
 import { AxiosResponse } from 'axios'
 import * as rax from 'retry-axios'
 
-import { BuildTradeInput, SupportedChainIds, SwapError, Trade } from '../../..'
+import { SupportedChainIds } from '@shapeshiftoss/types'
+import { BuildTradeInput, SwapError, Trade } from '../../..'
 import { ZrxQuoteResponse } from '../types'
 import { erc20AllowanceAbi } from '../utils/abi/erc20Allowance-abi'
 import { applyAxiosRetry } from '../utils/applyAxiosRetry'
